fixed bug with client gaps and borders
Previously, if a client was the only client on a monitor and got moved
to another monitor, it would still have its border and gap.

This was due to the logic in `resizeclient` using `selmon`, rather than
client's monitor. When a client is moved to another monitor, that
monitor does not get automagically selected, so this logic would always
fail.